+++ src/lib/Components/Inbox/Conversations/Previews/ArtworkPreview.tsx
@@ -3,12 +3,12 @@ import * as Relay from "react-relay"
 
 import { TouchableHighlight } from "react-native"
 
-import { PreviewText as P, Subtitle } from "../Typography"
+import { PreviewText as P, Subtitle } from "../../Typography"
 
 import styled from "styled-components/native"
-import colors from "../../../../data/colors"
-import fonts from "../../../../data/fonts"
-import OpaqueImageView from "../../OpaqueImageView"
+import colors from "../../../../../data/colors"
+import fonts from "../../../../../data/fonts"
+import OpaqueImageView from "../../../OpaqueImageView"
 
 const Container = styled.View`
   borderWidth: 1